Fix spec for `known_territory_subdivision` (#39)
`{:ok, nil}` is a valid response and not covered by the existing spec. Here I update the spec to account for this case.

There's an argument that maybe the library should return `{:ok, []}` when there are no known subdivisions for a valid territory. But as the `{:ok, nil}` was already the tested result I went with just updating the spec.